refactor: split up files and functions
This prepares for supporting better testing and future features
