+++ tests/go/testname_spec.lua
@@ -28,8 +28,9 @@ describe("Subtest name conversion", function()
   end)
 
   it("Special characters", function()
-    local expected_subtest_name = '"Comma , and \' are ok to use"'
-    local expected_gotest_name = "TestNames/Comma_,_and_'_are_ok_to_use"
+    local expected_subtest_name = '"Comma , and apostrophy \' are ok to use"'
+    local expected_gotest_name =
+      "TestNames/Comma_,_and_apostrophy_'_are_ok_to_use"
 
     -- Act
     local pos = tree:node(4):data()
